basetextoverlay: Do not take the BASE_TEXT_LOCK to negotiate caps
It can lead to deadlocks when flushed at the time we get are
negotiating srcpad caps. It looks like there is not reason why we were
holding the lock at that time.

Morever, in all code paths, that code is executed on serialized
events of the video sinkpad or the data flow so it is already protected
by the videosink stream lock.